Tests pass. Updated documentation for edge encoding.
Also removed the Project file in the test folder and tried to add them as  in the main Project file. The only complication came from adding  as a test-only dependency. I added the development version directly from url. However, when trying to remove SNaQ from  I would get an error when running tests saying that it was expecting SNaQ to be registered.

I think(?) this can be resolved by adding the main branch of SNaQ when those changes are finally pushed and registered.